requirements.txt
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
